refactor: use new exceptionhandler
